requirements.txt
openai==1.57.4
python-dotenv==1.0.0
pydantic==2.10.4
